test(#664): relocate env + wiki suites to integration (C6.1-spawn, dual-set)
env spawns a real `bash -c` to source the generated shell export script (real
shell env-sourcing boundary) and wiki spawns the real CLI; both also run
akmIndex({full:true}) in test bodies. They were grandfathered in BOTH
ALLOWED_SPAWN and ALLOWED_FULL_INDEX. Relocate out of the unit glob into
tests/integration/ (fix import depth + dynamic-import/__dirname CLI path),
drop both entries from BOTH sets (4 total), lowering UNIT_PURITY_BASELINE
23 -> 19, and repoint the wiki isolation-allowlist entry to the new path
(isolation baseline unchanged).
