+++ tests/integration/env.test.ts
@@ -3,14 +3,14 @@ import { spawnSync } from "node:child_process";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { buildShellExportScript, createEnv, injectIntoEnv, listKeys, loadEnv } from "../src/commands/env/env";
-import { getDbPath } from "../src/core/paths";
-import { closeDatabase, getAllEntries, openDatabase } from "../src/indexer/db/db";
-import { resetGraphBoostCache } from "../src/indexer/graph/graph-boost";
-import { akmIndex } from "../src/indexer/indexer";
-import { clearEmbeddingCache, resetLocalEmbedder } from "../src/llm/embedder";
-import { runCliCapture } from "./_helpers/cli";
-import { type Cleanup, sandboxStashDir, sandboxXdgCacheHome, sandboxXdgConfigHome, withEnv } from "./_helpers/sandbox";
+import { buildShellExportScript, createEnv, injectIntoEnv, listKeys, loadEnv } from "../../src/commands/env/env";
+import { getDbPath } from "../../src/core/paths";
+import { closeDatabase, getAllEntries, openDatabase } from "../../src/indexer/db/db";
+import { resetGraphBoostCache } from "../../src/indexer/graph/graph-boost";
+import { akmIndex } from "../../src/indexer/indexer";
+import { clearEmbeddingCache, resetLocalEmbedder } from "../../src/llm/embedder";
+import { runCliCapture } from "../_helpers/cli";
+import { type Cleanup, sandboxStashDir, sandboxXdgCacheHome, sandboxXdgConfigHome, withEnv } from "../_helpers/sandbox";
 
 // ── Test fixtures ───────────────────────────────────────────────────────────
