+++ tests/integration/wiki.test.ts
@@ -13,10 +13,10 @@ import fs from "node:fs";
 import http from "node:http";
 import os from "node:os";
 import path from "node:path";
-import { loadConfig, saveConfig } from "../src/core/config/config";
-import { akmIndex } from "../src/indexer/indexer";
-import { buildFileContext } from "../src/indexer/walk/file-context";
-import { wikiMatcher } from "../src/indexer/walk/matchers";
+import { loadConfig, saveConfig } from "../../src/core/config/config";
+import { akmIndex } from "../../src/indexer/indexer";
+import { buildFileContext } from "../../src/indexer/walk/file-context";
+import { wikiMatcher } from "../../src/indexer/walk/matchers";
 import {
   buildIngestWorkflow,
   createWiki,
@@ -38,11 +38,11 @@ import {
   stashRaw,
   validateWikiName,
   WIKIS_SUBDIR,
-} from "../src/wiki/wiki";
-import { type Cleanup, sandboxStashDir, sandboxXdgConfigHome } from "./_helpers/sandbox";
+} from "../../src/wiki/wiki";
+import { type Cleanup, sandboxStashDir, sandboxXdgConfigHome } from "../_helpers/sandbox";
 
 const tempDirs: string[] = [];
-const CLI = path.join(__dirname, "..", "src", "cli.ts");
+const CLI = path.join(__dirname, "..", "..", "src", "cli.ts");
 
 function makeStash(prefix = "akm-wiki-test-"): string {
   const dir = fs.mkdtempSync(path.join(os.tmpdir(), prefix));
@@ -375,7 +375,7 @@ describe("wikiMatcher", () => {
     const wikiDir = path.join(stash, WIKIS_SUBDIR, "research");
     fs.mkdirSync(wikiDir, { recursive: true });
     const abs = writePage(wikiDir, "SKILL.md", "# not actually a skill\n");
-    const { runMatchers } = await import("../src/indexer/walk/file-context");
+    const { runMatchers } = await import("../../src/indexer/walk/file-context");
     const ctx = buildFileContext(stash, abs);
     const result = await runMatchers(ctx);
     expect(result?.type).toBe("wiki");
